Move bundle RPC tests with implementation
